requirements.txt
flask
gpxpy
Pillow
piexif
anthropic
folium
python-dotenv
